menambahkan fungsi pemanggilan dan memperbaiki dashboard

+++ app/Http/Controllers/TransaksiController.php
@@ -72,6 +72,7 @@ class TransaksiController extends Controller
         $transaksi = Transaksi::with('pasien', 'obat', 'pemeriksaan')->findOrFail($id);
         return view('transaksi.show', compact('transaksi'));
     }
+    
 
     public function edit($id)
     {
@@ -84,34 +85,50 @@ class TransaksiController extends Controller
 
     public function update(Request $request, $id)
     {
+        $validator = Validator::make($request->all(), [
+            'id_pasien' => 'required|exists:pasien,id_pasien',
+            'tanggal_transaksi' => 'required|date', // Validasi untuk tanggal_transaksi
+            'id_obat' => 'required|array',
+            'id_obat.*' => 'required|exists:obat,id_obat',
+            'jumlah' => 'required|array',
+            'jumlah.*' => 'required|integer|min:1',
+            'id_jespem' => 'required|array',
+            'id_jespem.*' => 'required|exists:pemeriksaan,id_jespem',
+        ]);
+    
+        if ($validator->fails()) {
+            return redirect()->back()->withErrors($validator)->withInput();
+        }
+    
         $transaksi = Transaksi::findOrFail($id);
         $transaksi->update([
             'id_pasien' => $request->id_pasien,
             'tanggal_transaksi' => $request->tanggal_transaksi,
         ]);
-
+    
         $transaksi->obat()->detach();
         $transaksi->pemeriksaan()->detach();
-
+    
         $total_biaya = 0;
-
+    
         foreach ($request->id_obat as $index => $id_obat) {
             $jumlah = $request->jumlah[$index];
             $obat = Obat::find($id_obat);
             $total_biaya += $obat->harga * $jumlah;
             $transaksi->obat()->attach($id_obat, ['jumlah' => $jumlah]);
         }
-
+    
         foreach ($request->id_jespem as $id_jespem) {
             $pemeriksaan = Pemeriksaan::find($id_jespem);
             $total_biaya += $pemeriksaan->biaya;
             $transaksi->pemeriksaan()->attach($id_jespem);
         }
-
+    
         $transaksi->update(['total_biaya' => $total_biaya]);
-
+    
         return redirect()->route('transaksi.index')->with('success', 'Transaksi berhasil diperbarui');
     }
+    
 
     public function destroy($id)
     {
